+++ tests/test_config_load.py
@@ -15,12 +15,6 @@ def test_config_load_help():  # noqa: D103
     assert e.value.code == 0
 
 
-def test_config_load_switchver():  # noqa: D103
-    with pytest.raises(SystemExit) as e:
-        tfwrapper.main(["switchver"])
-    assert e.value.code == 2
-
-
 def test_config_load_init(tmp_working_dir_regional):  # noqa: D103
     paths = tmp_working_dir_regional
     os.chdir(paths["stack_dir"])
@@ -39,10 +33,18 @@ def test_config_load_init(tmp_working_dir_regional):  # noqa: D103
             terraform:
               vars:
                 myvar: myvalue
+                version: "1.0.3"
             """
         )
     )
     paths["terraform_conf"] = paths["stack_dir"] / "test.tf"
+    paths["terraform_conf"].write_text(
+        textwrap.dedent(
+            """
+            # Empty terraform configuration
+            """
+        )
+    )
 
     with pytest.raises(SystemExit) as e:
         tfwrapper.main(["init"])
@@ -67,10 +69,18 @@ def test_config_load_plan(tmp_working_dir_regional):  # noqa: D103
             terraform:
               vars:
                 myvar: myvalue
+                version: "1.0.3"
             """
         )
     )
     paths["terraform_conf"] = paths["stack_dir"] / "test.tf"
+    paths["terraform_conf"].write_text(
+        textwrap.dedent(
+            """
+            # Empty terraform configuration
+            """
+        )
+    )
 
     with pytest.raises(SystemExit) as e:
         tfwrapper.main(["plan"])